+++ apps/marginfi-v2-trading/src/components/common/search/components/index.ts
@@ -1,3 +1,3 @@
-export * from "./search-button";
 export * from "./search-drawer";
 export * from "./search-dialog";
+export * from "./search-list";
